Add files via upload
This foobar.vee is the working version from Bench 21.  If any of the benches' hp8757 loose their memory this utility can be used to reload the registers. It is the tech's responsibility to keep it updated and named appropriately. :) A readme file is provided in each UserObject for writing details of that LearnString's settings.